+++ lib/pages/task_pages/task_details_page.dart
@@ -1,5 +1,4 @@
 import 'dart:math';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:kronovo_app/pages/sub_task_pages/sub_task_details_page.dart';
@@ -8,6 +7,7 @@ import 'package:kronovo_app/pages/task_pages/update_task_page.dart';
 import '../../databases/sql_helper.dart';
 import '../../utils/responsive.dart';
 import '../../utils/theme.dart';
+import '../../widgets/confirmation_dialog.dart';
 import '../sub_task_pages/update_subtask_page.dart';
 
 class TaskDetailsPage extends StatefulWidget {
@@ -40,6 +40,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
   ];
   Random random = new Random();
 
+  // used fetch specific task data from task table using task id
   void _showTask(int id) async {
     if (id != null) {
       final data = await SQLHelper.getTask(id);
@@ -56,12 +57,27 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
     }
   }
 
+  // used to fetch all subtask record in subtask table which fall under specific task
+  // we fetch using task id
   void showSubTasks(int id) async {
     final data = await SQLHelper.getAllSubTasksByTask(id);
     setState(() {
       _listSubTasks = data;
     });
   }
+
+  final completeSnackBar = SnackBar(
+    content:   Row(
+      children: [
+        Icon(
+          Icons.celebration,
+          color: Colors.green,
+        ),Text('      Task is Completed', style: TextStyle(fontFamily: 'lato',color: Colors.black),),
+      ],
+    ),
+    duration: Duration(seconds: 3),
+    backgroundColor: Colors.white,
+  );
 
   @override
   void initState() {
@@ -72,12 +88,20 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
     });
   }
 
+  Future<void> _navigateforResult(BuildContext context) async {
+    final result = await Navigator.push(
+        context, MaterialPageRoute(builder: (context) =>
+        AddSubTaskPage(id: widget.id)));
+     if (!mounted) return;
+    initState();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(
-        title: Text('Task Details'),
+        title: Text('Task Details', style:  TextStyle(fontFamily: 'lato'),),
         centerTitle: true,
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.only(
@@ -89,11 +113,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
         actions: [
           IconButton(
               onPressed: () {
-                showDialog(
-                  builder: (context) => AddSubTaskPage(id: widget.id),
-                  context: context,
-                  barrierDismissible: false,
-                );
+                _navigateforResult(context);
               },
               icon: Icon(Icons.add))
         ],
@@ -180,7 +200,8 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                                   elevation: 2.0,
                                   label: Text(
                                     "${e.split(",").join()}",
-                                    style: TextStyle(
+                                    style:  TextStyle(
+                                      fontFamily: 'lato',
                                       fontSize: 16.0,
                                     ),
                                   ),
@@ -229,6 +250,8 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                             backgroundColor: MaterialStateProperty.all(Colors.blueAccent),
                             foregroundColor: MaterialStateProperty.all(Colors.white),
                             fixedSize: MaterialStateProperty.all(Size(wp(90, context), 50)),
+                            textStyle: MaterialStateProperty.all(
+                              TextStyle(fontFamily: 'lato',fontSize: 16,color: Colors.white),),
                             shape: MaterialStateProperty.all(
                               RoundedRectangleBorder(
                                 borderRadius: BorderRadius.circular(20),
@@ -242,16 +265,30 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                       Row(children: [
                         ElevatedButton(
                           onPressed: () {
-                            SQLHelper.deleteTask(widget.id);
-                            setState(() {
-                              Navigator.pop(context);
-                            });
+                            showDialog(
+                              context: context,
+                              builder: (BuildContext context) {
+                                return ConfirmationDialog(
+                                  title: 'Delete Project',
+                                  content: 'Are you sure you want to delete this Projet ?',
+                                  onConfirm: () {
+                                    SQLHelper.deleteTask(widget.id);
+                                    setState(() {
+                                      Navigator.pop(context);
+                                    });
+                                    Navigator.of(context).pop();
+                                  },
+                                );
+                              },
+                            );
                           },
                           child: const Text('DELETE TASK'),
                           style: ButtonStyle(
                             backgroundColor: MaterialStateProperty.all(Colors.red),
                             foregroundColor: MaterialStateProperty.all(Colors.white),
                             fixedSize: MaterialStateProperty.all(Size(wp(90, context), 50)),
+                            textStyle: MaterialStateProperty.all(
+                              TextStyle(fontFamily: 'lato',fontSize: 16,color: Colors.white),),
                             shape: MaterialStateProperty.all(
                               RoundedRectangleBorder(
                                 borderRadius: BorderRadius.circular(20),
@@ -272,7 +309,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                             padding: EdgeInsets.all(10),
                             child: Text(
                               'Sub Tasks',
-                              style:TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
+                              style:TextStyle(fontFamily: 'lato',fontSize: 18, fontWeight: FontWeight.bold),
                             ),
                           ),
                         ],
@@ -284,7 +321,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                 ),
               ),
               Container(
-                  child: ListView.builder(
+                  child: _listSubTasks.isNotEmpty ? ListView.builder(
                     primary: false,
                     scrollDirection: Axis.vertical,
                     shrinkWrap: true,
@@ -306,6 +343,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                             counter += 1;
                             SQLHelper.updateProgressTask(widget.id, progress,counter);
                             SQLHelper.changeValuesSubTask(_listSubTasks[index]['column_subtasks_id'], 1);
+                            ScaffoldMessenger.of(context).showSnackBar(completeSnackBar);
                             setState(() {
                               showSubTasks(widget.id);
                             });
@@ -349,6 +387,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                               Text(
                                 '${_listSubTasks[index]['subtasks_name']}',
                                 style: TextStyle(
+                                  fontFamily: 'lato',
                                   color: Colors.white,
                                   fontSize: 24,
                                   fontWeight: FontWeight.bold,
@@ -361,6 +400,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                                   Text(
                                     'Deadline :  ${_listSubTasks[index]['subtasks_end_date']}',
                                     style: TextStyle(
+                                      fontFamily: 'lato',
                                       color: Colors.white.withOpacity(0.8),
                                     ),
                                   ),
@@ -373,7 +413,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                                     height: 20,
                                     child: LinearProgressIndicator(
                                       value: _listSubTasks[index]['subtasks_progress'] /
-                                          _listSubTasks.length,
+                                          _listSubTasks.length.toDouble(),
                                       valueColor: AlwaysStoppedAnimation<Color>(
                                         Colors.white,
                                       ),
@@ -389,6 +429,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                                           '${(_listSubTasks[index]['subtasks_progress'] /
                                               _listSubTasks.length * 100).toInt()}%',
                                           style: TextStyle(
+                                            fontFamily: 'lato',
                                             color: Colors.white,
                                             fontWeight: FontWeight.bold,
                                           ),
@@ -406,7 +447,7 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                                   Text(
                                     '${convertToAgo(DateTime.parse(
                                         _listSubTasks[index]['createdAt']))}',
-                                    style: TextStyle(fontSize: 12.0, color: Colors.white),
+                                    style: TextStyle(fontFamily: 'lato',fontSize: 12.0, color: Colors.white),
                                   )
                                 ],
                               )
@@ -417,17 +458,31 @@ class _TaskDetailsPageState extends State<TaskDetailsPage> {
                     ),
                   ),
                 ),
-              )),
+              ) : Container(
+                    alignment: Alignment.center,
+                    margin: EdgeInsets.only(top: hp(3, context)),
+                    child: Column(
+                      children: <Widget>[
+                        Padding(
+                          padding: const EdgeInsets.only(left: 20.0),
+                          child: Image.asset("assets/images/default_icon.png", height: 100,width: 100),
+                        ),
+                        SizedBox(height: 6,),
+                        Text("NO SUB TASKS", style: TextStyle(fontFamily: 'lato'),),
+                        Text("Click On + to add Sub Task", style: TextStyle(fontFamily: 'lato'),),
+                      ],
+                    ),
+                  ),
+              ),
             ],
           ),
         ),
       ),
     );
   }
-
+  // this method is used to convert raw time data to (1 day ago) type data.
   String convertToAgo(DateTime input) {
     Duration diff = DateTime.now().difference(input);
-
     if (diff.inDays > 365)
       return "${(diff.inDays / 365).floor()} ${(diff.inDays / 365).floor() == 1 ? "year" : "years"} ago   ";
     if (diff.inDays > 30)
